requirements.txt
fpdf2>=2.7.0
peewee>=3.16.0